resolving jasmine adding some specs for LL

+++ src/util/LinkedList.js
@@ -1,5 +1,7 @@
 goog.provide('animatejs.util.LinkedList');
 
+goog.require('animatejs.util');
+goog.require('animatejs.util.LinkedListElement');
 goog.require('goog.Disposable');
 
 
@@ -73,19 +75,31 @@ animatejs.util.LinkedList.prototype.getTail = function() {
  * @param {animatejs.util.LinkedListElement=} opt_before
  * @protected
  */
-animatejs.util.LinkedList.prototype.addElement = function(element, opt_before) {
+animatejs.util.LinkedList.prototype.link = function(element, opt_before) {
   'use strict';
+  if (!animatejs.util.instanceOf(element, animatejs.util.LinkedListElement)) {
+    throw new TypeError();
+  }
+
+  if (opt_before &&
+      !animatejs.util.instanceOf(opt_before, animatejs.util.LinkedListElement)) {
+    throw new TypeError();
+  }
+
   var ownerList = element.getOwnerList();
   if (ownerList) {
     ownerList.unlink(element);
   }
 
-  if (opt_before) {
+  if (opt_before && this.hasElement(opt_before)) {
     element['prev'] = opt_before['prev'];
     element['next'] = opt_before;
     opt_before['prev'] = element;
   } else {
     element['prev'] = this.head_;
+    if (this.head_) {
+      this.head_['next'] = element;
+    }
     this.head_ = element;
   }
 
@@ -106,28 +120,52 @@ animatejs.util.LinkedList.prototype.addElement = function(element, opt_before) {
 animatejs.util.LinkedList.prototype.unlink = function(element) {
   'use strict';
 
+  if (!animatejs.util.instanceOf(element, animatejs.util.LinkedListElement)) {
+    throw new TypeError();
+  }
+
   var ownerList = element.getOwnerList();
   if (ownerList !== this) {
+    //unlink from the owning list
     ownerList.unlink(element);
     return;
   }
 
   if (element['prev'] && element['next']) {
+    //element from the middle of the list
     element['prev']['next'] = element['next'];
     element['next']['prev'] = element['prev'];
   } else if (element['prev']) {
+    //head element prev only
     element['prev']['next'] = null;
     this.head_ = element['prev'];
   } else if (element['next']) {
+    //tail element next only
     element['next']['prev'] = null;
     this.tail_ = element['next'];
   } else {
+    //no next prev the only element on the list
     this.head_ = null;
     this.tail_ = null;
   }
 
   element.setOwnerList(null);
   this.length_--;
+};
+
+
+/**
+ * Function checks if element is on the list
+ * @param {animatejs.util.LinkedListElement} element
+ * @return {boolean}
+ * @protected
+ */
+animatejs.util.LinkedList.prototype.hasElement = function(element) {
+  'use strict';
+  if (!animatejs.util.instanceOf(element, animatejs.util.LinkedListElement)) {
+    throw new TypeError();
+  }
+  return element.getOwnerList() === this;
 };
